Se hicieron las pruebas de la ultima modificación y se corrigieron los errores Se agrego una clase que maneja las excepciones del DeliveryException para devolver el mensaje de error (lo esta devolviendo en string)

+++ src/main/java/com/bd/tpfinal/controller/DeliveryController.java
@@ -251,7 +251,7 @@ public class DeliveryController {
     }
 
     @GetMapping(path = "/supplier/qualification/hasAtLeast/{stars}")
-    public List<Supplier>  getSupplierByQualificationValue(@PathVariable Integer stars){
+    public List<Supplier>  getSupplierByQualificationValue(@PathVariable Float stars){
         return this.service.getSupplierByQualificationValue(stars);
     }
